requirements.txt
sentence-transformers
faiss-cpu
PyPDF2
python-docx